refactor(web-app): refactor form models (#228)
close #228
