+++ apps/web-app/src/pages/Login/models/form.model.ts
@@ -1,4 +1,5 @@
-import { CInputUser, type TFormInputArray } from '@/models/UI'
+import { type TFormInputArray } from '@/components/Form/models'
+import { CInputUser } from '@/models/UI'
 import { getDynamicSchemas } from '@/utils'
 import { z as zValidator } from 'zod'
