+++ apps/web-app/src/pages/Settings/models/UI/form.model.ts
@@ -1,4 +1,5 @@
-import { CInputUser, type TFormInputArray } from '@/models/UI'
+import { type TFormInputArray } from '@/components/Form/models'
+import { CInputUser } from '@/models/UI'
 import { getDynamicSchemas } from '@/utils'
 import { z as ZValidator } from 'zod'
